+++ src/context/authContext/useAuth.ts
@@ -1,0 +1,4 @@
+import { useContext } from 'react';
+import { AuthUserContext } from './AuthUserContext';
+
+export const useAuth = () => useContext(AuthUserContext);
